+++ RELEASE_NOTES.md
@@ -6,6 +6,7 @@
     * [SER-3259] - External Service: Add new page for payment provider configuration
     * [SER-3364] - Implement label translation in Groups and Loans related UIs
     * [SER-3313] - Add bulk account transaction transfer
+    * [SER-3288] - External Service: Add new pages for order integration configuration
 
 ## Version 1.3.5 - Community 1.0.0